Add files via upload
Includes all features that are delivered via API on Glassnode. All features which satisfy two conditions are put appended to a dataframe. Firstly, that there must be sufficient data for that feature and secondly there must be relatively recent data present for that feature. From here, many univariate and multivariate models have been tested in relation to price forecasting for the following day, with a final ensemble model of 3 different algorithms all used to conduct theoretical trades on the last 100 days of the data to see if there is any improvement over the buy and hold strategy.